Move withings routes into a separate file.
